Add abstract content translation model
