+++ integreat_cms/api/tests/expected-outputs/augsburg_de_events.json
@@ -1,8 +1,8 @@
 [
     {
         "id": 1,
-        "url": "https://integreat.app/augsburg/de/events/test-veranstaltung",
-        "path": "/augsburg/de/events/test-veranstaltung",
+        "url": "http://localhost:8000/augsburg/de/events/test-veranstaltung/",
+        "path": "/augsburg/de/events/test-veranstaltung/",
         "title": "Test-Veranstaltung",
         "modified_gmt": "2020-01-22 12:46:33",
         "excerpt": "<p>Dies ist die Beschreibung einer Test-Veranstaltung.</p>",
@@ -10,7 +10,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/events/test-event"
+                "url": "http://localhost:8000/augsburg/en/events/test-event/",
+                "path": "/augsburg/en/events/test-event/"
             }
         },
         "thumbnail": null,
@@ -29,8 +30,8 @@
     },
     {
         "id": null,
-        "url": "https://integreat.app/augsburg/de/events/test-veranstaltung-2030-01-03",
-        "path": "/augsburg/de/events/test-veranstaltung-2030-01-03",
+        "url": "http://localhost:8000/augsburg/de/events/test-veranstaltung-2030-01-03/",
+        "path": "/augsburg/de/events/test-veranstaltung-2030-01-03/",
         "title": "Test-Veranstaltung",
         "modified_gmt": "2020-01-22 12:46:33",
         "excerpt": "<p>Dies ist die Beschreibung einer Test-Veranstaltung.</p>",
@@ -38,7 +39,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/events/test-event"
+                "url": "http://localhost:8000/augsburg/en/events/test-event/",
+                "path": "/augsburg/en/events/test-event/"
             }
         },
         "thumbnail": null,
@@ -57,8 +59,8 @@
     },
     {
         "id": null,
-        "url": "https://integreat.app/augsburg/de/events/test-veranstaltung-2030-01-15",
-        "path": "/augsburg/de/events/test-veranstaltung-2030-01-15",
+        "url": "http://localhost:8000/augsburg/de/events/test-veranstaltung-2030-01-15/",
+        "path": "/augsburg/de/events/test-veranstaltung-2030-01-15/",
         "title": "Test-Veranstaltung",
         "modified_gmt": "2020-01-22 12:46:33",
         "excerpt": "<p>Dies ist die Beschreibung einer Test-Veranstaltung.</p>",
@@ -66,7 +68,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/events/test-event"
+                "url": "http://localhost:8000/augsburg/en/events/test-event/",
+                "path": "/augsburg/en/events/test-event/"
             }
         },
         "thumbnail": null,
@@ -85,8 +88,8 @@
     },
     {
         "id": null,
-        "url": "https://integreat.app/augsburg/de/events/test-veranstaltung-2030-01-17",
-        "path": "/augsburg/de/events/test-veranstaltung-2030-01-17",
+        "url": "http://localhost:8000/augsburg/de/events/test-veranstaltung-2030-01-17/",
+        "path": "/augsburg/de/events/test-veranstaltung-2030-01-17/",
         "title": "Test-Veranstaltung",
         "modified_gmt": "2020-01-22 12:46:33",
         "excerpt": "<p>Dies ist die Beschreibung einer Test-Veranstaltung.</p>",
@@ -94,7 +97,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/events/test-event"
+                "url": "http://localhost:8000/augsburg/en/events/test-event/",
+                "path": "/augsburg/en/events/test-event/"
             }
         },
         "thumbnail": null,
@@ -113,8 +117,8 @@
     },
     {
         "id": null,
-        "url": "https://integreat.app/augsburg/de/events/test-veranstaltung-2030-01-29",
-        "path": "/augsburg/de/events/test-veranstaltung-2030-01-29",
+        "url": "http://localhost:8000/augsburg/de/events/test-veranstaltung-2030-01-29/",
+        "path": "/augsburg/de/events/test-veranstaltung-2030-01-29/",
         "title": "Test-Veranstaltung",
         "modified_gmt": "2020-01-22 12:46:33",
         "excerpt": "<p>Dies ist die Beschreibung einer Test-Veranstaltung.</p>",
@@ -122,7 +126,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/events/test-event"
+                "url": "http://localhost:8000/augsburg/en/events/test-event/",
+                "path": "/augsburg/en/events/test-event/"
             }
         },
         "thumbnail": null,
@@ -141,8 +146,8 @@
     },
     {
         "id": null,
-        "url": "https://integreat.app/augsburg/de/events/test-veranstaltung-2030-01-31",
-        "path": "/augsburg/de/events/test-veranstaltung-2030-01-31",
+        "url": "http://localhost:8000/augsburg/de/events/test-veranstaltung-2030-01-31/",
+        "path": "/augsburg/de/events/test-veranstaltung-2030-01-31/",
         "title": "Test-Veranstaltung",
         "modified_gmt": "2020-01-22 12:46:33",
         "excerpt": "<p>Dies ist die Beschreibung einer Test-Veranstaltung.</p>",
@@ -150,7 +155,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/events/test-event"
+                "url": "http://localhost:8000/augsburg/en/events/test-event/",
+                "path": "/augsburg/en/events/test-event/"
             }
         },
         "thumbnail": null,
