+++ integreat_cms/api/tests/expected-outputs/augsburg_de_locations.json
@@ -1,8 +1,8 @@
 [
     {
         "id": 1,
-        "url": "https://integreat.app/augsburg/de/pois/test-ort",
-        "path": "/augsburg/de/pois/test-ort",
+        "url": "http://localhost:8000/augsburg/de/pois/test-ort/",
+        "path": "/augsburg/de/pois/test-ort/",
         "title": "Test-Ort",
         "modified_gmt": "2020-01-22 12:52:13",
         "excerpt": "Kurzbeschreibung des Test-Ortes",
@@ -10,7 +10,8 @@
         "available_languages": {
             "en": {
                 "id": 2,
-                "url": "augsburg/en/pois/test-location"
+                "url": "http://localhost:8000/augsburg/en/pois/test-location/",
+                "path": "/augsburg/en/pois/test-location/"
             }
         },
         "thumbnail": null,
